# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.ChatPartnerResponseSwagger:
    properties:
      last_message:
        example: Apakah barang masih tersedia?
        type: string
      last_time:
        example: "2023-05-15T14:30:45Z"
        type: string
      partner_id:
        example: 2
        type: integer
      partner_nama:
        example: John Doe
        type: string
      unread:
        example: 3
        type: integer
    type: object
  domain.ChatResponseSwagger:
    properties:
      barang:
        $ref: '#/definitions/domain.ItemResponseSwagger'
      barang_id:
        example: 5
        type: integer
      created_at:
        example: "2023-05-15T14:30:45Z"
        type: string
      dibaca:
        example: false
        type: boolean
      id:
        example: 1
        type: integer
      penerima:
        $ref: '#/definitions/domain.UserResponseSwagger'
      penerima_id:
        example: 2
        type: integer
      pengirim:
        $ref: '#/definitions/domain.UserResponseSwagger'
      pengirim_id:
        example: 1
        type: integer
      pesan:
        example: Apakah barang masih tersedia?
        type: string
    type: object
...
  domain.ItemResponseSwagger:
    properties:
      deskripsi:
        example: Laptop gaming dengan spesifikasi tinggi
        type: string
      harga:
        example: 15000000
        type: integer
      id:
        example: 5
        type: integer
      kategori:
        example: Elektronik
        type: string
      nama:
        example: Laptop Asus ROG
        type: string
      penjual:
        $ref: '#/definitions/domain.UserResponseSwagger'
      penjual_id:
        example: 3
        type: integer
      status:
        example: Tersedia
        type: string
      view_url:
        example: http://endpoint.com/storage/buckets/bucket-id/files/file-id/view?project=project-id
        type: string
    type: object
...
  domain.SendMessageRequestSwagger:
    properties:
      barang_id:
        example: 5
        type: integer
      penerima_id:
        example: 2
        type: integer
      pesan:
        example: Apakah barang masih tersedia?
        type: string
    required:
    - barang_id
    - penerima_id
    - pesan
    type: object
...
  domain.UserResponseSwagger:
    properties:
      alamat:
        example: Jl. Contoh No. 123, Jakarta
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 2
        type: integer
      nama:
        example: John Doe
        type: string
      no_hp:
        example: "081234567890"
        type: string
      role:
        example: user
        type: string
    type: object
  utils.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/utils.Meta'
      success:
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@jubel.app
    name: Jubel API Support
  description: API untuk aplikasi jual beli barang bekas Jubel
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Jubel API
  version: "1.0"
paths:
...
  /chats:
    post:
      consumes:
      - application/json
      description: Mengirim pesan chat baru
      parameters:
      - description: Message data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.SendMessageRequestSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatResponseSwagger'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Send message
      tags:
      - chats
  /chats/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus pesan chat berdasarkan ID
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Delete chat
      tags:
      - chats
    get:
      consumes:
      - application/json
      description: Mendapatkan detail chat berdasarkan ID
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.ChatResponseSwagger'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get chat by ID
      tags:
      - chats
  /chats/{id}/read:
    patch:
      consumes:
      - application/json
      description: Menandai pesan chat sebagai telah dibaca
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Mark message as read
      tags:
      - chats
  /chats/barang/{id}:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar chat yang terkait dengan barang tertentu
      parameters:
      - description: Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ChatResponseSwagger'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get chats by item ID
      tags:
      - chats
  /chats/conversation:
    get:
      consumes:
      - application/json
      description: Mendapatkan percakapan antara dua pengguna tentang barang tertentu
      parameters:
      - description: Partner user ID
        in: query
        name: partner_id
        required: true
        type: integer
      - description: Item ID
        in: query
        name: barang_id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ChatResponseSwagger'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get conversation
      tags:
      - chats
  /chats/partners:
    get:
      consumes:
      - application/json
      description: Mendapatkan daftar pengguna yang pernah berkomunikasi dengan pengguna
        saat ini
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.ChatPartnerResponseSwagger'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.StandardResponse'
      security:
      - BearerAuth: []
      summary: Get chat partners
      tags:
      - chats
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"